+++ R/add_capital_distance.R
@@ -2,8 +2,8 @@
 #'
 #' @description `add_capital_distance()` allows you to add capital-to-capital
 #' distance to a (dyad-year, state-year) data frame. The distance variable that
-#' emerges (`capdist`) is calculated using the "Vincenty" method (i.e. "as the crow
-#' flies") and is expressed in kilometers.
+#' emerges (`capdist`) is calculated using the "Vincenty" method (i.e. "as the
+#' crow flies") and is expressed in kilometers.
 #'
 #' @return
 #'
@@ -15,7 +15,8 @@
 #' @details
 #'
 #' The function leans on attributes of the data that are provided by one of the
-#' "create" functions in this package (e.g. `create_dyadyear()` or `create_stateyear()`).
+#' "create" functions in this package (e.g. [create_dyadyears()] or
+#' [create_stateyear()]).
 #'
 #' Be advised that "jan1" and "dec31" are alternate specifications for
 #' "first" and "last" respectively and exist as kind of a nudge for what you